+++ components/Layout/Header/index.js
@@ -1,10 +1,21 @@
+import Gdpr from './Gdpr';
+import { GdprConsentConsumer } from '../../../context';
 import Logo from './Logo';
 import Nav from './Nav';
 import React from 'react';
 
+const getHeaderStyle = hasGdprConsent => (hasGdprConsent ? { top: '0px' } : {});
+
 export default ({ routes }) => (
-  <header className="Header">
-    <Logo />
-    <Nav routes={routes} />
-  </header>
+  <>
+    <Gdpr />
+    <GdprConsentConsumer>
+      {({ hasGdprConsent }) => (
+        <header className="Header" style={getHeaderStyle(hasGdprConsent)}>
+          <Logo />
+          <Nav routes={routes} />
+        </header>
+      )}
+    </GdprConsentConsumer>
+  </>
 );