+++ components/primitives/Callout.js
@@ -1,6 +1,6 @@
 import { func, object, string } from 'prop-types';
 
-import { Analytics } from './Analytics';
+import { AnalyticsConsumer } from '../../context';
 import { DOWNLOAD_PATH } from '../../common/constants';
 import React from 'react';
 import { withRouter } from 'next/router';
@@ -9,12 +9,9 @@ const Component = ({
   text = 'Download',
   className = '',
   router,
-  onClick = (e, track) => {
-    router.push(DOWNLOAD_PATH);
-    track.pageview(DOWNLOAD_PATH);
-  },
+  onClick = e => router.push(DOWNLOAD_PATH),
 }) => (
-  <Analytics>
+  <AnalyticsConsumer>
     {track => (
       <button
         onClick={e => {
@@ -33,7 +30,7 @@ const Component = ({
         {text}
       </button>
     )}
-  </Analytics>
+  </AnalyticsConsumer>
 );
 
 Component.propTypes = {