+++ components/primitives/NavLink.js
@@ -1,24 +1,15 @@
 import { shape, string } from 'prop-types';
 
-import { Analytics } from './Analytics';
 import NextLink from 'next/link';
 import React from 'react';
 import { withRouter } from 'next/router';
 
 const Link = ({ href, name, router: { pathname } }) => (
-  <Analytics>
-    {track => (
-      <NextLink href={href}>
-        <a
-          className={href === pathname ? 'active' : ''}
-          href={href}
-          onClick={() => track.pageview(href).send()}
-        >
-          {name}
-        </a>
-      </NextLink>
-    )}
-  </Analytics>
+  <NextLink href={href}>
+    <a className={href === pathname ? 'active' : ''} href={href}>
+      {name}
+    </a>
+  </NextLink>
 );
 
 Link.propTypes = {