+++ context/Analytics/index.js
@@ -1,20 +1,19 @@
 import React, { Component, createContext } from 'react';
 import { UA_ID, __DEV__ } from '../../common/config';
-
-import { node } from 'prop-types';
+import { node, string } from 'prop-types';
 
 const defaultApi = {
   event: ({ ec, ea, el, ev, dp }) => ({
-    send: () => console.table([{ ec, ea, el, ev, dp }]),
+    send: () => __DEV__ && console.table([{ ec, ea, el, ev, dp }]),
   }),
   pageview: path => ({
-    send: () => console.log(`Tracking pageview ${path}`),
+    send: () => __DEV__ && console.log(`Tracking pageview ${path}`),
   }),
 };
-const { Provider, Consumer } = createContext(defaultApi);
+const { Provider, Consumer: AnalyticsConsumer } = createContext(defaultApi);
 
 /**
- * Provides the universal analytics API to consumers.
+ * Provides the universal analytics API to consumers and tracks page views.
  *
  */
 export default class AnalyticsProvider extends Component {
@@ -23,13 +22,27 @@ export default class AnalyticsProvider extends Component {
   };
 
   /**
+   * Tracks page views once the API has been set.
+   *
+   * @return {null}
+   */
+  static getDerivedStateFromProps({ pathname }, { api }) {
+    api._translateParams && api.pageview(pathname).send();
+
+    return null;
+  }
+
+  /**
    * Sets the correct API to pass to consumers if we are client side.
+   * Ensures that IP addresses are anonymously sent to Google.
    *
    */
   componentDidMount() {
     const ua = require('universal-analytics');
     const { uid } = this;
     const api = ua(UA_ID, { uid });
+
+    api.set('anonymizeIp', true);
 
     this.setState({ api });
   }
@@ -59,6 +72,7 @@ export default class AnalyticsProvider extends Component {
 
 AnalyticsProvider.propTypes = {
   children: node,
+  pathname: string,
 };
 
-export const Analytics = Consumer;
+export { AnalyticsConsumer };