+++ pages/_app.js
@@ -1,21 +1,39 @@
 import '../style/style.scss';
 
-import { AnalyticsProvider, Layout } from '../components';
+import { AnalyticsProvider, GdprConsentProvider } from '../context';
 
 import App from 'next/app';
+import { COOKIE_KEY_GDPR_CONSENT } from '../common/constants';
+import { Layout } from '../components';
 import React from 'react';
+import Router from 'next/router';
+import nextCookies from 'next-cookies';
 import routes from '../content/routes';
 
 class MyApp extends App {
+  static async getInitialProps({ ctx }) {
+    const hasGdprConsent = nextCookies(ctx)[COOKIE_KEY_GDPR_CONSENT] || false;
+
+    return {
+      hasGdprConsent:
+        typeof hasGdprConsent === 'string'
+          ? JSON.parse(hasGdprConsent)
+          : hasGdprConsent,
+    };
+  }
+
   render() {
-    const { Component, pageProps, analytics } = this.props;
+    const { Component, ...pageProps } = this.props;
+    const pathname = Router.router ? Router.router.pathname : undefined;
 
     return (
-      <AnalyticsProvider api={analytics}>
-        <Layout routes={routes}>
-          <Component {...pageProps} />
-        </Layout>
-      </AnalyticsProvider>
+      <GdprConsentProvider hasGdprConsent={pageProps.hasGdprConsent}>
+        <AnalyticsProvider pathname={pathname}>
+          <Layout routes={routes}>
+            <Component {...pageProps} />
+          </Layout>
+        </AnalyticsProvider>
+      </GdprConsentProvider>
     );
   }
 }