+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## `v1.2.0` - 2019-10-16
+
+### Added
+
+- GDPR consent banner and modular styling to support it
+- GDPR context with a provider and consumer
+- Better analytics interop architecture
+- Cookie management packages
+
 ## `v1.1.1` - 2019-10-06
 
 ### Added